Use local inventory

+++ files/run.sh
@@ -19,7 +19,6 @@ fi
 source /etc/tower/conf.d/environment.sh
 
 ANSIBLE_REMOTE_TEMP=/tmp ANSIBLE_LOCAL_TEMP=/tmp ansible -i "127.0.0.1," -c local -v -m wait_for -a "host=$DATABASE_HOST port=$DATABASE_PORT" all
-ANSIBLE_REMOTE_TEMP=/tmp ANSIBLE_LOCAL_TEMP=/tmp ansible -i "127.0.0.1," -c local -v -m wait_for -a "host=$MEMCACHED_HOST port=$MEMCACHED_PORT" all
 ANSIBLE_REMOTE_TEMP=/tmp ANSIBLE_LOCAL_TEMP=/tmp ansible -i "127.0.0.1," -c local -v -m wait_for -a "host=$REDIS_HOST port=$REDIS_PORT" all
 
 if [[ $HOSTNAME == "awx" ]]; then
@@ -47,7 +46,12 @@ if [[ $HOSTNAME == "awx" ]]; then
   awx-manage register_queue --queuename=tower --instance_percent=100
 
   $(awx --conf.host http://awx-web:8052 login -f human)
-  cat /etc/supervisor_initialize.conf | sed "s/##TOWER_OAUTH_TOKEN##/${TOWER_OAUTH_TOKEN}/g" | tee -a /etc/supervisord_task.conf
+
+  if [[ ! -e /tmp/.supervisor_initialized ]]; then
+      cat /etc/supervisor_initialize.conf | sed "s/##TOWER_OAUTH_TOKEN##/${TOWER_OAUTH_TOKEN}/g" | tee -a /etc/supervisord_task.conf
+      cat /etc/supervisor_crond.conf | tee -a /etc/supervisord_task.conf
+      touch /tmp/.supervisor_initialized
+  fi
 
 else
 
@@ -59,6 +63,7 @@ unset $(cut -d = -f -1 /etc/tower/conf.d/environment.sh)
 
 if [[ $HOSTNAME == "awx" ]]; then
 
+  bash -c /rsync.sh
   exec supervisord -c /etc/supervisord_task.conf
 
 else
